Extract MockWebViewMessageHandlerDelegate to its own file (PR-5, WAL-10127)
Move the inline MockWebViewMessageHandlerDelegate class out of
WebViewMessageHandlerTests.swift and into a dedicated file at
Tests/WebTests/Mocks/MockWebViewMessageHandlerDelegate.swift.

Also documents why WalletTests and any future CrossmintClientTests each
maintain their own MockSmartWalletService rather than sharing one.
